Skip stboot build on aarch64

+++ stboot/build.sh
@@ -8,11 +8,16 @@ cp -fv ../userland/build/initramfs src/initramfs
 # Copy Secure Boot keys from sbkeys build
 cp -fv ../sbkeys/keys/* keys/
 
-# Build System Transparency bootloader and image
-podman build -t stboot .
+# TODO: Skip stboot build if we're on aarch64 since the
+#       build-stboot process seems to use x86_64-specific
+#       EFI code and fails on aarch64
+if [ "$(uname -m)" == "x86_64" ]; then
+    # Build System Transparency bootloader and image
+    podman build -t stboot .
 
-# Run container to copy stboot UKI, OS image, and
-# JSON manifest  into output dir
-podman run --rm -it -v "$(pwd)/build:/opt/out:Z" --entrypoint cp stboot -fv /opt/stboot.uki /opt/out/stboot.uki
-podman run --rm -it -v "$(pwd)/build:/opt/out:Z" --entrypoint cp stboot -fv /opt/nekos.json /opt/out/nekos.json
-podman run --rm -it -v "$(pwd)/build:/opt/out:Z" --entrypoint cp stboot -fv /opt/nekos.zip /opt/out/nekos.zip
+    # Run container to copy stboot UKI, OS image, and
+    # JSON manifest  into output dir
+    podman run --rm -it -v "$(pwd)/build:/opt/out:Z" --entrypoint cp stboot -fv /opt/stboot.uki /opt/out/stboot.uki
+    podman run --rm -it -v "$(pwd)/build:/opt/out:Z" --entrypoint cp stboot -fv /opt/nekos.json /opt/out/nekos.json
+    podman run --rm -it -v "$(pwd)/build:/opt/out:Z" --entrypoint cp stboot -fv /opt/nekos.zip /opt/out/nekos.zip
+fi
